Trimmed. Vol I.
