+++ godns.conf
@@ -27,7 +27,7 @@ password =""
 [log]
 stdout = true
 file = "./godns.log"
-level = "INFO"  #DEBUG | INFO |NOTICE | WARN | ERROR  
+level = "DEBUG"  #DEBUG | INFO |NOTICE | WARN | ERROR
 
 
 
@@ -52,7 +52,7 @@ refresh-interval = 5 # 5 seconds
 # Exclusive usage of a particular list of DNS recursive resolvers
 use-exclusively = true
 # List of DNS recursive resolvers
-backend-recursive-resolvers = [ "127.0.0.1:5353" ]
+backend-recursive-resolvers = [ "8.8.8.8:53" ]
 # Maximum response time of the Core backend found O.K. In ms.
 fit-response-time = 200 #ms
 # Delay between dry Core backend calls. In s.
