+++ handler.go
@@ -2,9 +2,9 @@ package main
 
 import (
 	"net"
-	"time"
 
 	"github.com/miekg/dns"
+	"time"
 )
 
 const (
@@ -24,9 +24,8 @@ func (q *Question) String() string {
 }
 
 type GODNSHandler struct {
-	resolver        *Resolver
-	cache, negCache Cache
-	hosts           Hosts
+	resolver      *Resolver
+	oraculumCache Cache
 }
 
 func NewHandler() *GODNSHandler {
@@ -35,7 +34,7 @@ func NewHandler() *GODNSHandler {
 		clientConfig    *dns.ClientConfig
 		cacheConfig     CacheSettings
 		resolver        *Resolver
-		cache, negCache Cache
+		oraculumCache   Cache
 	)
 
 	resolvConfig := settings.ResolvConfig
@@ -50,36 +49,19 @@ func NewHandler() *GODNSHandler {
 
 	cacheConfig = settings.Cache
 	switch cacheConfig.Backend {
+	// TODO might have other implementations...
 	case "memory":
-		cache = &MemoryCache{
-			Backend:  make(map[string]Mesg, cacheConfig.Maxcount),
+		oraculumCache = &MemoryCache{
+			Backend:  make(map[string]Data),
 			Expire:   time.Duration(cacheConfig.Expire) * time.Second,
 			Maxcount: cacheConfig.Maxcount,
 		}
-		negCache = &MemoryCache{
-			Backend:  make(map[string]Mesg),
-			Expire:   time.Duration(cacheConfig.Expire) * time.Second / 2,
-			Maxcount: cacheConfig.Maxcount,
-		}
-	case "redis":
-		// cache = &MemoryCache{
-		// 	Backend:    make(map[string]*dns.Msg),
-		//  Expire:   time.Duration(cacheConfig.Expire) * time.Second,
-		// 	Serializer: new(JsonSerializer),
-		// 	Maxcount:   cacheConfig.Maxcount,
-		// }
-		panic("Redis cache backend not implement yet")
 	default:
 		logger.Error("Invalid cache backend %s", cacheConfig.Backend)
 		panic("Invalid cache backend")
 	}
 
-	var hosts Hosts
-	if settings.Hosts.Enable {
-		hosts = NewHosts(settings.Hosts, settings.Redis)
-	}
-
-	return &GODNSHandler{resolver, cache, negCache, hosts}
+	return &GODNSHandler{resolver, oraculumCache}
 }
 
 func (h *GODNSHandler) do(Net string, w dns.ResponseWriter, req *dns.Msg) {
@@ -92,53 +74,16 @@ func (h *GODNSHandler) do(Net string, w dns.ResponseWriter, req *dns.Msg) {
 	} else {
 		remote = w.RemoteAddr().(*net.UDPAddr).IP
 	}
-	logger.Debug("%s lookup　%s", remote, Q.String())
+	logger.Debug("%s lookup %s", remote, Q.String())
 	logger.Debug("Question: %s", Q.String())
 	logger.Debug("w.RemoteAddr().String(): %s", w.RemoteAddr().String())
 
-	IPQuery := h.isIPQuery(q)
+	//karm cache
+	//IPQuery := h.isIPQuery(q)
 
-	// Query hosts
-	if settings.Hosts.Enable && IPQuery > 0 {
-		if ips, ok := h.hosts.Get(Q.qname, IPQuery); ok {
-			m := new(dns.Msg)
-			m.SetReply(req)
-
-			switch IPQuery {
-			case _IP4Query:
-				rr_header := dns.RR_Header{
-					Name:   q.Name,
-					Rrtype: dns.TypeA,
-					Class:  dns.ClassINET,
-					Ttl:    settings.Hosts.TTL,
-				}
-				for _, ip := range ips {
-					a := &dns.A{rr_header, ip}
-					m.Answer = append(m.Answer, a)
-				}
-			case _IP6Query:
-				rr_header := dns.RR_Header{
-					Name:   q.Name,
-					Rrtype: dns.TypeAAAA,
-					Class:  dns.ClassINET,
-					Ttl:    settings.Hosts.TTL,
-				}
-				for _, ip := range ips {
-					aaaa := &dns.AAAA{rr_header, ip}
-					m.Answer = append(m.Answer, aaaa)
-				}
-			}
-
-			w.WriteMsg(m)
-			logger.Debug("%s found in hosts file", Q.qname)
-			return
-		} else {
-			logger.Debug("%s didn't found in hosts file", Q.qname)
-		}
-	}
-
+	//karm cache
 	// Only query cache when qtype == 'A'|'AAAA' , qclass == 'IN'
-	key := KeyGen(Q)
+	/*key := KeyGen(Q)
 	if IPQuery > 0 {
 		mesg, err := h.cache.Get(key)
 		if err != nil {
@@ -157,30 +102,32 @@ func (h *GODNSHandler) do(Net string, w dns.ResponseWriter, req *dns.Msg) {
 			w.WriteMsg(&msg)
 			return
 		}
-	}
+	}*/
 
-	mesg, err := h.resolver.Lookup(Net, req, w.RemoteAddr())
+	mesg, err := h.resolver.Lookup(Net, req, w.RemoteAddr(), h.oraculumCache)
 
 	if err != nil {
 		logger.Warn("Resolve query error %s", err)
 		dns.HandleFailed(w, req)
 
 		// cache the failure, too!
+		/*karm cache
 		if err = h.negCache.Set(key, nil); err != nil {
 			logger.Warn("Set %s negative cache failed: %v", Q.String(), err)
-		}
+		}*/
 		return
 	}
 
 	w.WriteMsg(mesg)
 
+	/*karm cache
 	if IPQuery > 0 && len(mesg.Answer) > 0 {
 		err = h.cache.Set(key, mesg)
 		if err != nil {
 			logger.Warn("Set %s cache failed: %s", Q.String(), err.Error())
 		}
 		logger.Debug("Insert %s into cache", Q.String())
-	}
+	}*/
 }
 
 func (h *GODNSHandler) DoTCP(w dns.ResponseWriter, req *dns.Msg) {
